checkout now works!
what a glorious moment! I was stumped over this for over 12 hours.
there appear to be some latency involve while the javascript attempts
to collects the values from the already rendered html document. I had
to create a pop up dialog to confirm the checkout process so to give
the javascript enough time to populate all the data. Ridiculous!!