minor updates
gl web app is working properly now. I’ve just discovered that our
localhost server is a lot more relax on file naming convention, that
when I uploaded the files to the gl the codes doesn’t work! it turns
out that Main.php and main.php are not the same thing on gl while it is
ok locally. it took me over an hour to debut the mistake! check your
type case everyone. use camel case and lowercase where necessary.