more updates with many bugs fixed and new bugs created
the waitlist has a bug when the user selects to show their position in
line.
when the user selects to wait for a plane at an airport, it will report
them as waiting at multiple airport, depending on how many airport has
that plane on their waiting list.
This is due to the waitlist implementation issue and not the code
fault.